backend/pyproject.toml
[project]
name = "moviebackend"
version = "0.1.0"
description = ""
authors = [
    {name = "elshafieeltayeb",email = "elshafie.eltayeb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (==2.0.40)",
    "alembic (>=1.15.2,<2.0.0)",
    "ruff (>=0.11.9,<0.12.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "injector (>=0.22.0,<0.23.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "shell (>=1.0.1,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "repository" },
    { include = "api" },
    { include = "domain_models" }
]
